+++ example-project/src/test/java/some/template/CucableJavaTemplate.java
@@ -3,8 +3,8 @@ package some.template;
 import cucumber.api.CucumberOptions;
 
 @CucumberOptions(
-        features = {"target/parallel/features/[FEATURE_FILE_NAME].feature"},
-        plugin = {"json:target/cucumber-report/[FEATURE_FILE_NAME].json"}
+        features = {"target/parallel/features/[CUCABLE:FEATURE].feature"},
+        plugin = {"json:target/cucumber-report/[CUCABLE:RUNNER].json"}
 )
 public class CucableJavaTemplate {