+++ plugin-code/src/main/java/com/trivago/rta/features/FeatureFileConverter.java
@@ -24,15 +24,17 @@ import com.trivago.rta.gherkin.GherkinDocumentParser;
 import com.trivago.rta.logging.CucableLogger;
 import com.trivago.rta.properties.PropertyManager;
 import com.trivago.rta.runners.RunnerFileContentRenderer;
+import com.trivago.rta.vo.FeatureRunner;
 import com.trivago.rta.vo.SingleScenario;
-import com.trivago.rta.vo.SingleScenarioRunner;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 
 import static com.trivago.rta.logging.CucableLogger.CucableLogLevel.COMPACT;
 import static com.trivago.rta.logging.CucableLogger.CucableLogLevel.DEFAULT;
@@ -86,16 +88,21 @@ public class FeatureFileConverter {
      * @param featureFilePaths feature files to process
      * @throws CucablePluginException see {@link CucablePluginException}
      */
-    public void convertToSingleScenariosAndRunners(
+    public void generateSingleScenarioFeatures(
             final List<Path> featureFilePaths) throws CucablePluginException {
-        int scenarioCounter = 0;
+        int featureFileCounter = 0;
+        List<String> allGeneratedFeaturePaths = new ArrayList<>();
+
         for (Path featureFilePath : featureFilePaths) {
-            scenarioCounter += convertToSingleScenariosAndRunners(featureFilePath);
+            List<String> generatedFeatureFilePaths = generateSingleScenarioFeatures(featureFilePath);
+            allGeneratedFeaturePaths.addAll(generatedFeatureFilePaths);
+            featureFileCounter += generatedFeatureFilePaths.size();
         }
+        int runnerFileCounter = generateRunnerClasses(allGeneratedFeaturePaths, propertyManager.getDesiredNumberOfRunners());
         logger.info("-------------------------------------", DEFAULT);
         logger.info(
-                String.format("Cucable created %d separate scenario(s) from the provided feature(s).", scenarioCounter),
-                DEFAULT, COMPACT, MINIMAL
+                String.format("Cucable created %d separate feature file(s) and %d runner(s) from the provided feature(s).",
+                        featureFileCounter, runnerFileCounter), DEFAULT, COMPACT, MINIMAL
         );
     }
 
@@ -103,14 +110,13 @@ public class FeatureFileConverter {
      * Converts all scenarios in the given feature file to single
      * scenario feature files and their respective runners.
      *
-     * @param featureFilePath feature file to process.
+     * @param sourceFeatureFilePath feature file to process.
      * @return Number of created scenarios.
      * @throws CucablePluginException see {@link CucablePluginException}
      */
-    private int convertToSingleScenariosAndRunners(final Path featureFilePath)
+    private List<String> generateSingleScenarioFeatures(final Path sourceFeatureFilePath)
             throws CucablePluginException {
-
-        String featureFilePathString = featureFilePath.toString();
+        String featureFilePathString = sourceFeatureFilePath.toString();
 
         if (featureFilePathString == null || featureFilePathString.equals("")) {
             throw new MissingFileException(featureFilePathString);
@@ -141,9 +147,12 @@ public class FeatureFileConverter {
             throw new CucablePluginException("There is no parsable scenario or scenario outline at line " + lineNumbers);
         }
 
+        // Stores all generated feature file names and associated source feature paths for later runner creation
+        List<String> generatedFeaturePaths = new ArrayList<>();
+
         for (SingleScenario singleScenario : singleScenarios) {
             String renderedFeatureFileContent = featureFileContentRenderer.getRenderedFeatureFileContent(singleScenario);
-            String featureFileName = getFeatureFileNameFromPath(featureFilePath);
+            String featureFileName = getFeatureFileNameFromPath(sourceFeatureFilePath);
             Integer featureCounter = singleFeatureCounters.getOrDefault(featureFileName, 0);
             featureCounter++;
             String scenarioCounterFilenamePart = String.format(SCENARIO_COUNTER_FORMAT, featureCounter);
@@ -169,26 +178,87 @@ public class FeatureFileConverter {
                 // Save scenario information to new feature file
                 fileIO.writeContentToFile(renderedFeatureFileContent, generatedFeatureFilePath);
 
-                // Generate runner for the newly generated single scenario feature file
-                SingleScenarioRunner singleScenarioRunner =
-                        new SingleScenarioRunner(
-                                propertyManager.getSourceRunnerTemplateFile(), generatedFileName);
-
-                String renderedRunnerFileContent =
-                        runnerFileContentRenderer.getRenderedRunnerFileContent(singleScenarioRunner, singleScenario);
-
-                String generatedRunnerFilePath =
-                        propertyManager.getGeneratedRunnerDirectory()
-                                .concat(PATH_SEPARATOR)
-                                .concat(generatedFileName)
-                                .concat(RUNNER_FILE_EXTENSION);
-
-                fileIO.writeContentToFile(renderedRunnerFileContent, generatedRunnerFilePath);
+                generatedFeaturePaths.add(generatedFileName);
             }
         }
-        int createdScenarios = singleScenarios.size();
-        logProcessCompleteMessage(featureFilePathString, createdScenarios);
-        return createdScenarios;
+        logFeatureFileConversionMessage(featureFilePathString, singleScenarios.size());
+        return generatedFeaturePaths;
+    }
+
+    /**
+     * Generate runner classes for a list of feature file paths.
+     *
+     * @param generatedFeatureNames  The list of generated feature file names.
+     * @param numberOfDesiredRunners The number of desired runners (if set to 0, a runner is generated for each feature file path).
+     * @return The number of generated runners.
+     * @throws CucablePluginException see {@link CucablePluginException}.
+     */
+    private int generateRunnerClasses(final List<String> generatedFeatureNames, final int numberOfDesiredRunners) throws CucablePluginException {
+
+        int targetRunnerNumber = numberOfDesiredRunners;
+        if (targetRunnerNumber == 0) {
+            targetRunnerNumber = generatedFeatureNames.size();
+        }
+
+        List<List<String>> generatedFeatureNamesPerRunner = new ArrayList<>(targetRunnerNumber);
+        for (int i = 0; i < targetRunnerNumber; i++) {
+            generatedFeatureNamesPerRunner.add(new ArrayList<>());
+        }
+
+        int currentRunnerIndex = 0;
+        for (String generatedFeatureName : generatedFeatureNames) {
+            generatedFeatureNamesPerRunner.get(currentRunnerIndex).add(generatedFeatureName);
+            currentRunnerIndex++;
+            if (currentRunnerIndex >= targetRunnerNumber) {
+                currentRunnerIndex = 0;
+            }
+        }
+
+        int runnerFileCounter = 0;
+        for (List<String> generatedFeatureNamesForSingleRunner : generatedFeatureNamesPerRunner) {
+            if (generatedFeatureNamesForSingleRunner.size() > 0) {
+                generateRunnerClass(generatedFeatureNamesForSingleRunner);
+                runnerFileCounter++;
+            }
+        }
+
+        return runnerFileCounter;
+    }
+
+    /**
+     * Generate a single runner class from a list of feature files.
+     *
+     * @param generatedFeatureFileNames The list of generated generated feature file names.
+     * @throws CucablePluginException see {@link CucablePluginException}.
+     */
+    private void generateRunnerClass(final List<String> generatedFeatureFileNames) throws CucablePluginException {
+
+        // The runner class name will be equal to the feature name if there is only one feature to run.
+        // Otherwise, a generated runner class name is used.
+        String runnerClassName;
+        if (generatedFeatureFileNames.size() == 1) {
+            runnerClassName = generatedFeatureFileNames.get(0);
+        } else {
+            runnerClassName = "CucableMultiRunner_"
+                    .concat(UUID.randomUUID().toString().replace("-", "_"))
+                    .concat(INTEGRATION_TEST_POSTFIX);
+        }
+
+        // Generate runner for the newly generated single scenario feature file
+        FeatureRunner featureRunner =
+                new FeatureRunner(
+                        propertyManager.getSourceRunnerTemplateFile(), runnerClassName, generatedFeatureFileNames);
+
+        String renderedRunnerClassContent =
+                runnerFileContentRenderer.getRenderedRunnerFileContent(featureRunner);
+
+        String generatedRunnerClassFilePath =
+                propertyManager.getGeneratedRunnerDirectory()
+                        .concat(PATH_SEPARATOR)
+                        .concat(runnerClassName)
+                        .concat(RUNNER_FILE_EXTENSION);
+
+        fileIO.writeContentToFile(renderedRunnerClassContent, generatedRunnerClassFilePath);
     }
 
     /**
@@ -197,12 +267,12 @@ public class FeatureFileConverter {
      * @param featureFileName  The name of the processed feature file.
      * @param createdScenarios The number of created scenarios for the feature file.
      */
-    private void logProcessCompleteMessage(String featureFileName, final int createdScenarios) {
+    private void logFeatureFileConversionMessage(String featureFileName, final int createdScenarios) {
         String logPostfix = ".";
         if (propertyManager.hasValidScenarioLineNumbers()) {
             logPostfix = String.format(" with line number(s) %s.", propertyManager.getScenarioLineNumbers());
         }
-        logger.info(String.format("- %3d scenario(s) from %s%s", createdScenarios, featureFileName, logPostfix),
+        logger.info(String.format("- %3d feature(s) <= %s%s", createdScenarios, featureFileName, logPostfix),
                 DEFAULT);
     }