+++ plugin-code/src/main/java/com/trivago/rta/vo/Step.java
@@ -42,13 +42,4 @@ public class Step {
     public String getDocString() {
         return docString;
     }
-
-    @Override
-    public String toString() {
-        return "Step{" +
-                "name='" + name + '\'' +
-                ", dataTable=" + dataTable +
-                ", docString='" + docString + '\'' +
-                '}';
-    }
 }